+++ MessageApp/clientMessageApp.js
@@ -1,48 +1,23 @@
-const express = require('express');
-const app = express();
-const axios = require('axios');
-const bodyParser = require('body-parser')
-import MessageApp from '../service/index'
+import axios from 'axios';
 
+class MessageApp {
+    constructor() {
+        this.service = axios.create({
+            baseURL: `http://localhost:9001`
+        })
+    }
 
-const validator = (req, res, next) => {
-    let { destination, body } = req.body;
-    if (typeof destination != String || typeof body != String) {
-        console.log("Error")
-        res.send("Incorrect")
-    }
-    else if (destination == "" || body == "") {
-        console.log("Error")
-        res.send("Destination or body are empty")
-    }
-    else if (destination.length > 20 || body.length > 240) {
-        console.log("Error")
-        res.send("Destination or body are too long")
-    }
-    else if (destination == null || body == null) {
-        console.log("Error")
-        res.send("Destination or body are incorrect")
-    }
-    else if (destination == undefined || body == undefined) {
-        console.log("Error")
-        res.send("Destination or body are incorrect")
-    } else next();
-
-
-    app.post('/message', validator(), (req, res, next) => {
-        let { destination, body } = req.body;
-        axios.post('http://messageapp:3000/message', { destination, body })
-            .then(resp => {
-                res.status(200).
-                    res.send(`${resp.data}`)
-            })
-            .catch(e => {
-                res.status(500).
-                    console.log(e)
-            })
-    })
+sendMessage = (destination, body) => {
+    return this.service.post('/message', {destination, body})
+    .then(response => response.data)
+    .catch(error => error.response)
 }
 
-app.listen(9001, () => {
-    console.log('listend on port 9001!');
-});
+getMessages = () => {
+    return this.service.get('/messages')
+    .then(response => response.data)
+    .catch(error => error.response)
+  }
+}
+
+export default MessageApp
