+++ docker-compose.yml
@@ -4,6 +4,9 @@ services:
     build: .
     ports:
      - "9003:9003"
+     depends_on:
+      - mongodb
+    command: ["mongodb:27027","-t","5", "--","node", "app.js"]
   messageapp:
     image: cabify/backend-bootcamp-messageapp
     ports:
@@ -13,4 +16,4 @@ services:
     volumes:
       - ./data/db:/data/db
     ports:
-      - 27017:27017
+      - "27017:27017"
