--- package.json
@@ -1,15 +1,0 @@
-{
-  "name": "exercise1",
-  "version": "0.0.0",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "start": "index.js",
-    "dev": "DEBUG=exercise1:* nodemon"
-  },
-  "dependencies": {
-    "axios": "^0.18.0",
-    "express": "^4.16.4",
-    "mongoose": "^5.3.8"
-  }
-}
