+++ service/Dockerfile
@@ -2,4 +2,4 @@ FROM node:latest
 ADD . /app
 WORKDIR /app
 RUN npm install
-CMD ["node", "/service/index.js"]
+CMD ["node", "app.js"]
