+++ service/package.json
@@ -1,14 +1,16 @@
 {
-  "name": "exercise1",
-  "version": "0.0.0",
-  "private": true,
+  "name": "exersice1",
+  "version": "1.0.0",
+  "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "index.js",
-    "dev": "DEBUG=exercise1:* nodemon"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
+  "author": "",
+  "license": "ISC",
   "dependencies": {
     "axios": "^0.18.0",
+    "dotenv": "^6.1.0",
     "express": "^4.16.4",
     "mongoose": "^5.3.8"
   }
